Update generated code for tests.
All but 1 of the tests migrated from streams_old passes.

+++ streams/vocab/gen_property_accuracy_interface.go
@@ -19,7 +19,7 @@ type AccuracyPropertyInterface interface {
 	Clear()
 	// Get returns the value of this property. When IsFloat returns false, Get
 	// will return any arbitrary value.
-	Get() float32
+	Get() float64
 	// GetIRI returns the IRI of this property. When IsIRI returns false,
 	// GetIRI will return any arbitrary value.
 	GetIRI() *url.URL
@@ -53,7 +53,7 @@ type AccuracyPropertyInterface interface {
 	Serialize() (interface{}, error)
 	// Set sets the value of this property. Calling IsFloat afterwards will
 	// return true.
-	Set(v float32)
+	Set(v float64)
 	// SetIRI sets the value of this property. Calling IsIRI afterwards will
 	// return true.
 	SetIRI(v *url.URL)
